Common type for Distribution Status messages
