+++ Example/Source/Images.xcassets/ic_dfu.symbolset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  },
+  "symbols" : [
+    {
+      "filename" : "sync_sync_symbol.svg",
+      "idiom" : "universal"
+    }
+  ]
+}
